#4207 Fix File Structure
