+++ src/backend/src/prisma/seed/organization.process.ts
@@ -1,7 +1,7 @@
 // processes/organization.process.ts
 import { Organization } from '@prisma/client';
-import { SeedProcess } from '../seed-process.js';
-import { bootstrapUserCreateInput, organizationCreateInput } from '../../factories/organization.factory.js';
+import { SeedProcess } from '../processes/seed-process.js';
+import { bootstrapUserCreateInput, organizationCreateInput } from '../factories/organization.factory.js';
 
 export type OrganizationOutput = {
   organization: Organization;
